Render imgui on correct D3D9 render target (#15)
It looks like Guilty internally maintains two surfaces: one that corresponds to the "game" window, and one that corresponds to the "UI". The "game" window changes resolutions depending on where in the game it is, and during combat, it's running at Guilty's original native 4:3, with the viewport at the top left of the "game" window. We don't want to draw there- we want to draw on the "UI" surface, where the viewport is at the top left of the actual Windows window. This is the same surface that the sidebars is drawn on.

As an aside, the weird surface selection is probably why OBS doesn't capture +R correctly if you try to "game capture", but it does work if you "window capture".