Updated Cheattable & Imgui functionality (#18)
I added the "Enable Hitboxes" option to the CheatTable and restructured the Imgui implementation. The windows are built within the createImguiWindows() function and are called via a main menu window. In addition, i added the beginnings of a CheatTable window to mimic the layout and information found within the CheatTable file.