Add user feedback when the synchronization thread is running (#47)
Previously, once you clicked the "join" or "host" in the GGPO windows, the button was still available. On the host, pressing the button again would create two synchronization server threads, both of which would attempt to bind to the same address, crashing the game. On the client, while it wouldn't crash, pressing join could result in unexpected behavior. For example, two threads could both try to send synchronization formation to the same host but the host closes the connection after one of the threads sends synchronization info.

This diff replaces the button with (admittedly not fantastically-written) text when the synchronization thread has already started.