Save/load more RNG state, to prevent RNG desyncs (#59)
After further reversing, the RNG was bigger than I'd previously thought. Plus, there are other RNG buffers hanging out in the system. This diff increases the data stored and saves, loads, and transmits all RNG data. After trying to reproduce the RNG desyncs that I've previously seen with Faust using the save/load state debug tool for somewhere in the neighborhood of 10 minutes and not seeing one, I think this an upper bound on the RNG state we need to save and load.

As a bonus, I found that there could be an RNG desync caused by different loading times; the main RNG cursor increments every frame while not in a network match, including during the fade-out and fade-in between the versus interstitial and the actual gameplay that hides a loading screen. This hacks in similar logic for when the player's not in a GGPO match, which should prevent the RNG from incrementing during loads, as long as the conditional is correct.