Fix rollback during timesync (#64)
Crashes would occur if a client needed to rollback while it was also dropping frames to timesync because ggpo_advance_frame_callback called the _fake_ simulation. The fake simulation would recognize that it was timesyncing, not advance the frame, and then GGPO would detect a mismatch between the amount of frames that it tried to roll forward and the amount of frames it succeeded in rolling forward.