Fix flickers during GGPO input failures and during timesync (#66)
When GGPO input proccessing fails, we don't simulate the next state, but we do still attempt to render. This results in the render having nothing to render, creating a black screen. This is particularly noticeable in out-of-timesync or high-latency scenarios, where we may be hitting max-prediction-frames multiple times a second, causing multiple flickers.

Instead, make sure we also don't draw if we're skipping state simulation by checking that the lastResult was a success inside the begin- and end-scene methods.

This gives us a pretty good blueprint of what to do with GGPO timesync as well. We need to move the decrementing of our "frames-ahead" value until after the last piece of code that needs it- in this case, the end-scene method.