Add unlock-all and relock-all utilities (#68)
This means people don't need to download a 100% save file for development purposes. It's not really great for end-users, but it's okay enough for now.

You could get to this from the debug pause menu, but the debug pause menu is a little jank and annoying, so I added it to the topbar directly.