Add Linux support via Proton (#73)
FindGuilty() doesn't work on Linux, since the Steam location registry key doesn't exist in the wineprefix. To get around this, I made the launcher check for the GGPOPLUSR_LINUX_DIR environment variable and use that in place of calling FindGuilty(). I also included a scipt to automatically run all of this through Steam, and added instructions on how to do all of this to the readme.